refactor(Command.java): 객체 패키지 위치 변경
